DDF-1405 Changes github-to-bamboo hubot script to pick appropriate PR build plans based on file deltas
With the repositories merged, our PR builds will now provisionally build only the submodule changed and run the distro + integration tests builds. If a PR has changes that span multiple submodules, we will build the entire repository; likewise, any change to the root of the repository (the reactor pom, in most cases) will result in a full build.

Changes to the distribution submodule will only run a build of that single submodule.